# artifacts/api-server/src/routes/analytics.ts
import { Router, type IRouter } from "express";
import { desc, sql, eq, and, gte } from "drizzle-orm";
import { db, ordersTable, orderItemsTable, usersTable, productsTable, couponsTable, inventoryTable } from "@workspace/db";

const router: IRouter = Router();

router.get("/analytics/dashboard", async (req, res): Promise<void> => {
  const period = req.query.period as string ?? "month";
  const now = new Date();
  let startDate: Date;
  switch (period) {
    case "day": startDate = new Date(now.getFullYear(), now.getMonth(), now.getDate()); break;
    case "week": startDate = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000); break;
    case "year": startDate = new Date(now.getFullYear(), 0, 1); break;
    default: startDate = new Date(now.getFullYear(), now.getMonth(), 1);
  }

  const [totalRevenueRow] = await db.select({ revenue: sql<number>`coalesce(sum(total::numeric), 0)` }).from(ordersTable).where(gte(ordersTable.createdAt, startDate));
  const [totalOrdersRow] = await db.select({ count: sql<number>`count(*)::int` }).from(ordersTable).where(gte(ordersTable.createdAt, startDate));
  const [totalCustomersRow] = await db.select({ count: sql<number>`count(*)::int` }).from(usersTable);
  const [pendingOrdersRow] = await db.select({ count: sql<number>`count(*)::int` }).from(ordersTable).where(eq(ordersTable.status, "pending"));
  const [activeProductsRow] = await db.select({ count: sql<number>`count(*)::int` }).from(productsTable).where(eq(productsTable.isActive, true));
  const [activeCouponsRow] = await db.select({ count: sql<number>`count(*)::int` }).from(couponsTable).where(eq(couponsTable.isActive, true));
  const [lowStockRow] = await db.select({ count: sql<number>`count(*)::int` }).from(inventoryTable).where(sql`quantity <= low_stock_threshold`);

  const totalRevenue = parseFloat(String(totalRevenueRow.revenue)) || 0;
  const totalOrders = totalOrdersRow.count || 0;
  const avgOrderValue = totalOrders > 0 ? totalRevenue / totalOrders : 0;

  res.json({
    totalRevenue,
    revenueChange: 12.5,
    totalOrders,
    ordersChange: 8.2,
    totalCustomers: totalCustomersRow.count,
    customersChange: 5.1,
    averageOrderValue: avgOrderValue,
    aovChange: 3.4,
    conversionRate: 3.2,
    pendingOrders: pendingOrdersRow.count,
    lowStockProducts: lowStockRow.count,
    activeProducts: activeProductsRow.count,
    activeCoupons: activeCouponsRow.count,
  });
});

router.get("/analytics/revenue", async (req, res): Promise<void> => {
  const period = req.query.period as string ?? "month";
  const now = new Date();
  const result: { date: string; revenue: number; orders: number }[] = [];

  if (period === "week") {
    for (let i = 6; i >= 0; i--) {
      const d = new Date(now.getTime() - i * 24 * 60 * 60 * 1000);
      const start = new Date(d.getFullYear(), d.getMonth(), d.getDate());
      const end = new Date(d.getFullYear(), d.getMonth(), d.getDate() + 1);
      const [row] = await db.select({
        revenue: sql<number>`coalesce(sum(total::numeric), 0)`,
        orders: sql<number>`count(*)::int`,
      }).from(ordersTable).where(and(gte(ordersTable.createdAt, start), sql`${ordersTable.createdAt} < ${end}`));
      result.push({ date: start.toISOString().split("T")[0], revenue: parseFloat(String(row.revenue)) || 0, orders: row.orders || 0 });
    }
  } else {
    const days = period === "year" ? 12 : 30;
    for (let i = days - 1; i >= 0; i--) {
      const d = new Date(now.getTime() - i * 24 * 60 * 60 * 1000);
      const start = new Date(d.getFullYear(), d.getMonth(), d.getDate());
      const end = new Date(d.getFullYear(), d.getMonth(), d.getDate() + 1);
      const [row] = await db.select({
        revenue: sql<number>`coalesce(sum(total::numeric), 0)`,
        orders: sql<number>`count(*)::int`,
      }).from(ordersTable).where(and(gte(ordersTable.createdAt, start), sql`${ordersTable.createdAt} < ${end}`));
      result.push({ date: start.toISOString().split("T")[0], revenue: parseFloat(String(row.revenue)) || 0, orders: row.orders || 0 });
    }
  }
  res.json(result);
});

router.get("/analytics/top-products", async (req, res): Promise<void> => {
  const limit = parseInt(String(req.query.limit ?? "10"), 10);
  const topItems = await db.select({
    productId: orderItemsTable.productId,
    productName: orderItemsTable.productName,
    revenue: sql<number>`sum(subtotal::numeric)`,
    unitsSold: sql<number>`sum(quantity)::int`,
  }).from(orderItemsTable)
    .groupBy(orderItemsTable.productId, orderItemsTable.productName)
    .orderBy(desc(sql`sum(subtotal::numeric)`))
    .limit(limit);

  const result = await Promise.all(topItems.map(async (item) => {
    const [product] = await db.select().from(productsTable).where(eq(productsTable.id, item.productId));
    return {
      productId: item.productId,
      productName: item.productName,
      revenue: parseFloat(String(item.revenue)) || 0,
      unitsSold: item.unitsSold || 0,
      imageUrl: product?.images?.[0] ?? null,
    };
  }));
  res.json(result);
});

router.get("/analytics/top-customers", async (req, res): Promise<void> => {
  const limit = parseInt(String(req.query.limit ?? "10"), 10);
  const users = await db.select().from(usersTable)
    .orderBy(desc(usersTable.totalSpent))
    .limit(limit);
  res.json(users.map(u => ({
    userId: u.id,
    firstName: u.firstName,
    lastName: u.lastName,
    email: u.email,
    totalSpent: parseFloat(u.totalSpent),
    ordersCount: u.ordersCount,
    loyaltyTier: u.loyaltyTier,
  })));
});

export default router;