# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hebrew RTL E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ─────────────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/coupon:
    post:
      operationId: applyCouponToCart
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyCouponBody"
      responses:
        "200":
          description: Cart with coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeCouponFromCart
      tags: [cart]
      summary: Remove coupon from cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /analytics/dashboard:
    get:
      operationId: getAnalyticsDashboard
      tags: [analytics]
      summary: Get admin KPI dashboard data
      parameters:
        - name: period
          in: query
          schema:
            type: ["string", "null"]
            enum: [day, week, month, year, null]
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsDashboard"
  /analytics/revenue:
    get:
      operationId: getRevenueChart
      tags: [analytics]
      summary: Get revenue chart data
      parameters:
        - name: period
          in: query
          schema:
            type: ["string", "null"]
            enum: [week, month, year, null]
      responses:
        "200":
          description: Revenue chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /analytics/top-products:
    get:
      operationId: getTopProducts
      tags: [analytics]
      summary: Get top products by revenue
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProductStat"
  /analytics/top-customers:
    get:
      operationId: getTopCustomers
      tags: [analytics]
      summary: Get top customers by spend
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCustomerStat"

  # ─── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nameHe:
          type: string
        nameEn:
          type: ["string", "null"]
        slug:
          type: string
        descriptionHe:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        costPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        brandId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties: true
        stockQuantity:
          type: integer
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        viewsCount:
          type: integer
        salesCount:
          type: integer
        ratingAverage:
          type: number
        ratingCount:
          type: integer
        weight:
          type: ["number", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameHe
        - slug
        - price
        - stockQuantity
        - isActive
        - isFeatured
        - viewsCount
        - salesCount
        - ratingAverage
        - ratingCount
        - images
        - tags
        - createdAt
        - updatedAt
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, admin]
        loyaltyPoints:
          type: integer
        loyaltyTier:
          type: string
          enum: [bronze, silver, gold, vip]
        totalSpent:
          type: number
        ordersCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - loyaltyPoints
        - loyaltyTier
        - totalSpent
        - ordersCount
        - isActive
        - createdAt
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
      required:
        - email
        - password
        - firstName
        - lastName
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required:
        - productId
        - product
        - quantity
        - price
        - subtotal
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        shipping:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: number
        loyaltyPointsUsed:
          type: integer
        loyaltyDiscount:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - discount
        - shipping
        - total
        - couponDiscount
        - loyaltyPointsUsed
        - loyaltyDiscount
        - itemCount
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
    ApplyCouponBody:
      type: object
      properties:
        code:
          type: string
      required:
        - code

    # ── Order ──
...
    AnalyticsDashboard:
      type: object
      properties:
        totalRevenue:
          type: number
        revenueChange:
          type: number
        totalOrders:
          type: integer
        ordersChange:
          type: number
        totalCustomers:
          type: integer
        customersChange:
          type: number
        averageOrderValue:
          type: number
        aovChange:
          type: number
        conversionRate:
          type: number
        pendingOrders:
          type: integer
        lowStockProducts:
          type: integer
        activeProducts:
          type: integer
        activeCoupons:
          type: integer
      required:
        - totalRevenue
        - revenueChange
        - totalOrders
        - ordersChange
        - totalCustomers
        - customersChange
        - averageOrderValue
        - aovChange
        - conversionRate
        - pendingOrders
        - lowStockProducts
        - activeProducts
        - activeCoupons
    RevenueDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders
    TopProductStat:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        revenue:
          type: number
        unitsSold:
          type: integer
        imageUrl:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - revenue
        - unitsSold
    TopCustomerStat:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        totalSpent:
          type: number
        ordersCount:
          type: integer
        loyaltyTier:
          type: string
      required:
        - userId
        - firstName
        - lastName
        - email
        - totalSpent
        - ordersCount
        - loyaltyTier

    # ── Admin ──
...